Move snakes commands into fun folder
