+++ bot/resources/fun/snakes/snake_cards/card_top.png
