+++ bot/resources/fun/snakes/snake_facts.json
