+++ bot/resources/fun/snakes/snake_idioms.json
