+++ bot/resources/fun/snakes/snake_names.json
