+++ bot/resources/fun/snakes/snake_quiz.json
