+++ bot/resources/fun/snakes/snakes_and_ladders/banner.jpg
