+++ bot/resources/fun/snakes/special_snakes.json
